rename package to all lowercase name
